+++ src/providers/index.ts
@@ -1,5 +1,6 @@
 export { ReviewProvider } from './provider.js';
 export { GitLabProvider } from './gitlab/gitlab-provider.js';
 export { GitHubProvider } from './github/github-provider.js';
+export { BitbucketCloudProvider } from './bitbucket/bitbucket-cloud-provider.js';
 export { LocalGitProvider } from './local/local-git-provider.js';
 export { createProvider } from './factory.js';